+++ R/print.p_direction.R
@@ -5,14 +5,14 @@ print.p_direction <- function(x, digits = 2, ...) {
   } else if ("data.frame" %in% class(x)) {
     .print_pd(x, digits, ...)
   } else {
-    cat(sprintf("pd = %.*f%%", digits, x * 100))
+    cat(sprintf("pd = %s%%", insight::format_value(x * 100, digits = digits)))
   }
 }
 
-
+#' @keywords internal
 .print_pd <- function(x, digits, ...) {
   insight::print_color("# Probability of Direction (pd)\n\n", "blue")
   x$Parameter <- as.character(x$Parameter)
-  x$pd <- sprintf("%.*f%%", digits, x$pd * 100)
+  x$pd <- sprintf("%s%%", insight::format_value(x$pd * 100, digits = digits))
   print_data_frame(x, digits = digits)
 }