+++ NEWS.md
@@ -1,7 +1,8 @@
-# bayestestR 0.3.1
+# bayestestR 0.3.2
 
 ## New functions / features
 
+- `p_significance()` Compute the probability of Practical Significance (ps)
 - `overlap()` Compute overlap between two empirical distributions
 - `estimate_density()`: `method = "mixture"` argument added for mixture density estimation